+++ src/app/index.run.js
@@ -1,7 +1,7 @@
 import { Visualizer } from '@uirouter/visualizer';
+import { states as administrationStates } from './pages/administration/administration.state.js';
 import { states as collectionsStates } from './pages/collections/collections.state.js';
 import { states as dashboardStates } from './pages/dashboard/dashboard.state.js';
-import { states as listingStates } from './pages/listing/listing.state.js';
 
 (() => {
     'use strict';
@@ -21,6 +21,16 @@ import { states as listingStates } from './pages/listing/listing.state.js';
                     let needsRedirect = false;
 
                     // load states dependent on features
+                    if (featureFlags.isOn('change-request')) {
+                        administrationStates['change-request'].forEach(state => {
+                            if ($uiRouter.stateRegistry.get(state.name)) {
+                                $uiRouter.stateRegistry.deregister(state.name);
+                            }
+                            $uiRouter.stateRegistry.register(state);
+                            needsReload = needsReload || $state.$current.name === state.name;
+                        });
+                    }
+
                     if (featureFlags.isOn('role-developer')) {
                         dashboardStates['role-developer'].forEach(state => {
                             if ($uiRouter.stateRegistry.get(state.name)) {
@@ -33,24 +43,6 @@ import { states as listingStates } from './pages/listing/listing.state.js';
                         dashboardStates['role-developer'].forEach(state => {
                             $uiRouter.stateRegistry.deregister(state.name);
                             needsRedirect = needsRedirect || $state.$current.name === state.name;
-                        });
-                    }
-
-                    if (featureFlags.isOn('listing-edit')) {
-                        listingStates['listing-edit-on'].forEach(state => {
-                            if ($uiRouter.stateRegistry.get(state.name)) {
-                                $uiRouter.stateRegistry.deregister(state.name);
-                            }
-                            $uiRouter.stateRegistry.register(state);
-                            needsReload = needsReload || $state.$current.name === state.name;
-                        });
-                    } else {
-                        listingStates['listing-edit-off'].forEach(state => {
-                            if ($uiRouter.stateRegistry.get(state.name)) {
-                                $uiRouter.stateRegistry.deregister(state.name);
-                            }
-                            $uiRouter.stateRegistry.register(state);
-                            needsReload = needsReload || $state.$current.name === state.name;
                         });
                     }