+++ src/app/services/network.service.js
@@ -255,24 +255,16 @@ export class NetworkService {
         return this.getActivity(call, activityRange);
     }
 
-    getComplaints () {
-        return this.apiGET('/complaints');
+    getChangeRequests () {
+        return this.apiGET('/change-requests');
     }
 
-    getComplaintStatusTypes () {
-        return this.apiGET('/data/complaint-status-types');
+    getChangeRequestStatusTypes () {
+        return this.apiGET('/data/change-request-status-types');
     }
 
-    getComplainantTypes () {
-        return this.apiGET('/data/complainant-types');
-    }
-
-    getCriteria () {
-        return this.apiGET('/data/certification-criteria');
-    }
-
-    getCriterionProductStatistics () {
-        return this.apiGET('/statistics/criterion_product');
+    getChangeRequestTypes () {
+        return this.apiGET('/data/change-request-types');
     }
 
     getCmsDownload () {
@@ -308,6 +300,22 @@ export class NetworkService {
             return this.apiGET('/collections/developers');
             //no default
         }
+    }
+
+    getComplaints () {
+        return this.apiGET('/complaints');
+    }
+
+    getComplainantTypes () {
+        return this.apiGET('/data/complainant-types');
+    }
+
+    getCriteria () {
+        return this.apiGET('/data/certification-criteria');
+    }
+
+    getCriterionProductStatistics () {
+        return this.apiGET('/statistics/criterion_product');
     }
 
     getDeveloper (developerId) {
@@ -465,7 +473,7 @@ export class NetworkService {
     }
 
     getSearchOptions () {
-        return this.apiGET('/data/search_options');
+        return this.apiGET('/data/search-options');
     }
 
     getSedParticipantStatisticsCount () {
@@ -526,11 +534,11 @@ export class NetworkService {
             .then(function (response) {
                 data.nonconformityStatusTypes = response;
             });
-        this.apiGET('/data/surveillance_requirements')
+        this.apiGET('/data/surveillance-requirements')
             .then(function (response) {
                 data.surveillanceRequirements = response;
             });
-        this.apiGET('/data/nonconformity_types')
+        this.apiGET('/data/nonconformity-types')
             .then(function (response) {
                 data.nonconformityTypes = response;
             });
@@ -720,12 +728,20 @@ export class NetworkService {
         return this.apiPOST('/versions/' + versionObject.oldVersion.versionId + '/split', versionObject);
     }
 
+    submitChangeRequest (request) {
+        return this.apiPOST('/change-requests', request);
+    }
+
     unimpersonateUser () {
         return this.apiGET('/auth/unimpersonate');
     }
 
     updateAnnualSurveillanceReport (report) {
         return this.apiPUT('/surveillance-report/annual', report);
+    }
+
+    updateChangeRequest (changeRequest) {
+        return this.apiPUT('/change-requests', changeRequest)
     }
 
     updateComplaint (complaint) {